@drieseng 2021-11-23
This one is ready to go.

 I'll have to revisit some benchmarks where we are slower with new implementation because we put an enormous pressure on the .NET threadpool. These are benchmarks that I haven't even pushed yet, where I configured the threadpool to have a **MaxConcurrency** of **int.MaxValue**. Let's say that this is not your typical configuration :-)